Adding a init command for control channels

+++ storage/specialTextChannel/console/reloadConsole.js
@@ -1,16 +1,17 @@
 const config = require("../../config.json");
 const fs = require("fs");
-const path = config.location+"ConsoleFile.txt";
+const path = config.location + "ConsoleFile.txt";
 const idChannel = config.idConsoleChannel;
 
 
 module.exports.run = async (bot) => {
+    if (idChannel == null) return;
     const channel = await bot.channels.fetch(idChannel);
-    await bot.basicFunctions.get("deleteAll").run(bot,channel);
+    await bot.basicFunctions.get("deleteAll").run(bot, channel);
 
     fs.readFile(path, 'utf8', function (err, data) {
         if (err) throw err;
-        
+
         const dataSplit = data.split("\n");
         let msg = "";
         for (let index = 0; index < dataSplit.length; index++) {
